Moved Logo component out of NavigationBar and into separate file.
